Исправил ограничение "принадлежность точки линии", изменив уравнение. Теперь если при решении методом Гаусса на главной диагонали будет ноль (после перестановки уравнений) он будет заменён на единицу.

+++ NutonSolver/NutonSolver/includes/Matrix.cpp
@@ -3,12 +3,12 @@
 #include <stdio.h>
 
 //	
-void	Print_matrix(double** arr, int size)
+void	Print_matrix(double** arr, size_t size)
 {
-	for (int i = 0; i < size; i++)
+	for (size_t i = 0; i < size; i++)
 	{
-		for (int j = 0; j < size; j++)
-			printf("%.3f\t", arr[i][j]);
+		for (size_t j = 0; j < size; j++)
+			printf("%.2f\t", arr[i][j]);
 		printf("\n");
 	}
 }
